\name{GWmodelVis-package}
\alias{GWmodelVis-package}
\alias{GWmodelVis}
\docType{package}
\title{
Visualization Tools for Geographically Weighted Models
}
\description{
In 'GWmodelVis', we developed visualization tools for a branch of spatial statistics
termed geographically weighted (GW) models, including functions to calibrate:
GW summary statistics, GW principal components analysis, GW discriminant analysis,
and various forms of GW regression (provided in basic and robust forms).
Sonification techniques are adopted to embed result visualization from GW models.
To enable video rendering, users must install 'FFmpeg' (\url{https://ffmpeg.org/download.html}).
}
\details{
\tabular{ll}{
Package: \tab GWmodelVis\cr
Type: \tab Package\cr
Version: \tab 1.0-0\cr
Date: \tab 2025-03-31\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Binbin Lu, Huimei Wang

Maintainer: Binbin Lu <binbinlu@whu.edu.cn>
}
\note{
To enable video rendering in 'GWmodelVis', install 'FFmpeg' (\url{https://ffmpeg.org/download.html}).
Package examples and tests write output ONLY to \code{tempdir()} to comply with CRAN policies.
}
\references{
Gollini et al. (2015) <doi:10.18637/jss.v063.i17>

Lu et al. (2014) <doi:10.1080/10095020.2014.917453>

Lu et al. (2024) "Gwmodels: A standalone software..." <https://doi.org/10.1080/10095020.2024.2343011>

Lu et al. (2023) "Gwmodels: A software..." <https://doi.org/10.1016/j.softx.2022.101291>
}
\keyword{package}
